ref(dashboards): Address review feedback for dashboard revisions UI
- Switch useDashboardRevisions to useQuery + apiOptions (new preferred pattern)
- Move DashboardRevision type and query options factory to the hook file
  since they are only used there; export for future cache invalidation
- Replace custom styled table with SimpleTable core component
