+++ static/app/views/dashboards/hooks/useDashboardRevisions.tsx
@@ -1,25 +1,31 @@
-import {
-  makeDashboardRevisionsQueryKey,
-  type DashboardRevision,
-} from 'sentry/actionCreators/dashboards';
-import {useApiQuery} from 'sentry/utils/queryClient';
+import {useQuery} from '@tanstack/react-query';
+
+import {apiOptions} from 'sentry/utils/api/apiOptions';
 import {useOrganization} from 'sentry/utils/useOrganization';
+
+export type DashboardRevision = {
+  createdBy: {email: string; id: string; name: string} | null;
+  dateCreated: string;
+  id: string;
+  source: 'edit' | 'pre-restore';
+  title: string;
+};
+
+export function makeDashboardRevisionsQueryOptions(orgSlug: string, dashboardId: string) {
+  return apiOptions.as<DashboardRevision[]>()(
+    '/organizations/$organizationIdOrSlug/dashboards/$dashboardId/revisions/',
+    {
+      path: {organizationIdOrSlug: orgSlug, dashboardId},
+      staleTime: 30_000,
+    }
+  );
+}
 
 interface UseDashboardRevisionsOptions {
   dashboardId: string;
-  enabled?: boolean;
 }
 
-export function useDashboardRevisions({
-  dashboardId,
-  enabled = true,
-}: UseDashboardRevisionsOptions) {
+export function useDashboardRevisions({dashboardId}: UseDashboardRevisionsOptions) {
   const organization = useOrganization();
-  return useApiQuery<DashboardRevision[]>(
-    makeDashboardRevisionsQueryKey(organization.slug, dashboardId),
-    {
-      staleTime: 30_000,
-      enabled,
-    }
-  );
+  return useQuery(makeDashboardRevisionsQueryOptions(organization.slug, dashboardId));
 }
